feat: move js modules into a sub folder
We only want propshaft to transpile the application level js and not the individal modules, so I have moved the modules into their own folder.
