+++ app/javascript/application.js
@@ -1,7 +1,7 @@
 // Entry point for the build script in your package.json
 import initHeader from '@citizensadvice/design-system/lib/header';
-import initGreedyNav from "./greedy-nav";
-import initSupplierTableButton from "./supplier-table";
+import initGreedyNav from "./modules/greedy-nav";
+import initSupplierTableButton from "./modules/supplier-table";
 
 try {
   // Initialise design-system modules first
