+++ app/javascript/modules/last_added_appliance.js
