+++ app/javascript/modules/supplier-table.js
